Updates to tutorial code and added tutorial text explaining the code
