+++ CMakeLists.txt
@@ -10,6 +10,7 @@ find_package(Boost REQUIRED system filesystem date_time thread)
 find_package(ament_cmake REQUIRED)
 find_package(control_msgs REQUIRED)
 find_package(moveit_core REQUIRED)
+find_package(moveit_task_constructor_core REQUIRED)
 find_package(moveit_ros_planning REQUIRED)
 find_package(moveit_ros_planning_interface REQUIRED)
 find_package(moveit_ros_perception REQUIRED)
@@ -50,6 +51,7 @@ set(THIS_PACKAGE_INCLUDE_DEPENDS
   Boost
   control_msgs
   moveit_servo
+  moveit_task_constructor_core
 )
 
 include_directories(${THIS_PACKAGE_INCLUDE_DIRS})
