--- doc/tutorials/mtc_pick_and_place_tutorial/CMakeLists.txt
@@ -1,31 +1,0 @@
-cmake_minimum_required(VERSION 3.8)
-project(mtc_pick_and_place_tutorial)
-
-if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
-  add_compile_options(-Wall -Wextra -Wpedantic)
-endif()
-
-find_package(ament_cmake REQUIRED)
-find_package(rclcpp REQUIRED)
-find_package(moveit_core REQUIRED)
-find_package(moveit_task_constructor_core REQUIRED)
-
-set(THIS_PACKAGE_INCLUDE_DEPENDS
-    rclcpp
-    moveit_core
-    moveit_task_constructor_core
-    )
-
-add_executable(${PROJECT_NAME} src/main.cpp)
-ament_target_dependencies(${PROJECT_NAME} ${THIS_PACKAGE_INCLUDE_DEPENDS})
-
-install(TARGETS ${PROJECT_NAME}
-        ARCHIVE DESTINATION lib
-        LIBRARY DESTINATION lib
-        RUNTIME DESTINATION lib/${PROJECT_NAME})
-
-install(DIRECTORY launch
-  DESTINATION share/${PROJECT_NAME}
-)
-
-ament_package()
