--- doc/tutorials/mtc_pick_and_place_tutorial/src/main.cpp
@@ -1,277 +1,0 @@
-#include <rclcpp/rclcpp.hpp>
-#include <moveit/planning_scene/planning_scene.h>
-#include <moveit/planning_scene_interface/planning_scene_interface.h>
-#include <moveit/task_constructor/task.h>
-#include <moveit/task_constructor/solvers.h>
-#include <moveit/task_constructor/stages.h>
-#include <tf2_geometry_msgs/tf2_geometry_msgs.hpp>
-#include <tf2_eigen/tf2_eigen.hpp>
-
-static const rclcpp::Logger LOGGER = rclcpp::get_logger("mtc_tutorial");
-
-class MTCTaskNode
-{
-public:
-  MTCTaskNode(const rclcpp::NodeOptions& options);
-
-  rclcpp::node_interfaces::NodeBaseInterface::SharedPtr getNodeBaseInterface();
-
-  void doTask();
-
-  void setupPlanningScene();
-
-private:
-  // Compose an MTC task from a series of stages.
-  moveit::task_constructor::Task createTask();
-
-
-private:
-  rclcpp::Node::SharedPtr node_;
-};
-
-rclcpp::node_interfaces::NodeBaseInterface::SharedPtr MTCTaskNode::getNodeBaseInterface()
-{
-  return node_->get_node_base_interface();
-}
-
-MTCTaskNode::MTCTaskNode(const rclcpp::NodeOptions& options)
-: node_{ std::make_shared<rclcpp::Node>("mtc_node", options) }
-{
-
-}
-
-void MTCTaskNode::setupPlanningScene()
-{
-  moveit_msgs::msg::CollisionObject object;
-  object.id = "object";
-  object.header.frame_id = "world";
-  object.primitives.resize(1);
-  object.primitives[0].type = shape_msgs::msg::SolidPrimitive::CYLINDER;
-  object.primitives[0].dimensions = { 0.1, 0.02 };
-
-  geometry_msgs::msg::Pose pose;
-  pose.position.x = 0.5;
-  pose.position.y = -0.25;
-  object.pose = pose;
-
-  moveit::planning_interface::PlanningSceneInterface psi;
-  psi.applyCollisionObject(object);
-}
-
-
-void MTCTaskNode::doTask()
-{
-  moveit::task_constructor::Task task = createTask();
-
-  try
-  {
-    task.init();
-  }
-  catch (moveit::task_constructor::InitStageException& e) {
-
-    RCLCPP_ERROR_STREAM(LOGGER, e);
-    return;
-  }
-
-  if (!task.plan())
-  {
-    RCLCPP_ERROR_STREAM(LOGGER, "Task planning failed");
-    return;
-  }
-
-  auto result = task.execute(*task.solutions().front());
-  if (result.val != moveit_msgs::msg::MoveItErrorCodes::SUCCESS)
-  {
-    RCLCPP_ERROR_STREAM(LOGGER, "Task execution failed");
-    return;
-  }
-
-  return;
-}
-
-moveit::task_constructor::Task MTCTaskNode::createTask()
-{
-  moveit::task_constructor::Task task;
-  task.stages()->setName("demo task");
-  task.loadRobotModel(node_);
-
-  const auto& arm_group_name = "panda_arm";
-  const auto& eef_name = "hand";
-  const auto& hand_group_name = "hand";
-  const auto& hand_frame = "panda_hand";
-
-  // Set task properties
-  task.setProperty("group", arm_group_name);
-  task.setProperty("eef", eef_name);
-  task.setProperty("hand", hand_group_name);
-  task.setProperty("hand_grasping_frame", hand_frame);
-  task.setProperty("ik_frame", hand_frame);
-
-  moveit::task_constructor::Stage* current_state_ptr = nullptr;  // Forward current_state on to grasp pose generator
-  auto stage_state_current = std::make_unique<moveit::task_constructor::stages::CurrentState>("current");
-  current_state_ptr = stage_state_current.get();
-  task.add(std::move(stage_state_current));
-
-  auto sampling_planner = std::make_shared<moveit::task_constructor::solvers::PipelinePlanner>(node_);
-  auto interpolation_planner = std::make_shared<moveit::task_constructor::solvers::JointInterpolationPlanner>();
-
-  auto cartesian_planner = std::make_shared<moveit::task_constructor::solvers::CartesianPath>();
-  cartesian_planner->setMaxVelocityScaling(1.0);
-  cartesian_planner->setMaxAccelerationScaling(1.0);
-  cartesian_planner->setStepSize(.01);
-
-  auto scene{ std::make_shared<planning_scene::PlanningScene>(task.getRobotModel()) };
-  auto& robot_state{ scene->getCurrentStateNonConst() };
-
-//  auto initial{ std::make_unique<moveit::task_constructor::stages::FixedState>("start") };
-//  initial->setState(scene);
-//  task.add(std::move(initial));
-
-//  {
-//    auto connect{ std::make_unique<moveit::task_constructor::stages::Connect>(
-//       "path length", moveit::task_constructor::stages::Connect::GroupPlannerVector{ { "panda_arm", sampling_planner } }) };
-//                                                                 // demonstration purposes
-//    task.add(std::move(connect));
-//  }
-
-
-  auto stage_open_hand = std::make_unique<moveit::task_constructor::stages::MoveTo>("open hand", interpolation_planner);
-  stage_open_hand->setGroup(hand_group_name);
-  stage_open_hand->setGoal("open");
-  task.add(std::move(stage_open_hand));
-
-  auto stage_move_to_pick = std::make_unique<moveit::task_constructor::stages::Connect>(
-      "move to pick", moveit::task_constructor::stages::Connect::GroupPlannerVector{ { arm_group_name, sampling_planner } });
-  stage_move_to_pick->setTimeout(5.0);
-  stage_move_to_pick->properties().configureInitFrom(moveit::task_constructor::Stage::PARENT);
-  task.add(std::move(stage_move_to_pick));
-
-  moveit::task_constructor::Stage* attach_object_stage = nullptr;  // Forward attach_object_stage to place pose generator
-  {
-    auto grasp = std::make_unique<moveit::task_constructor::SerialContainer>("pick object");
-    task.properties().exposeTo(grasp->properties(), { "eef", "hand", "group", "ik_frame" });
-    grasp->properties().configureInitFrom(moveit::task_constructor::Stage::PARENT, { "eef", "hand", "group", "ik_frame" });
-
-    {
-      auto stage = std::make_unique<moveit::task_constructor::stages::MoveRelative>("approach object", cartesian_planner);
-      stage->properties().set("marker_ns", "approach_object");
-      stage->properties().set("link", hand_frame);
-      stage->properties().configureInitFrom(moveit::task_constructor::Stage::PARENT, { "group" });
-      stage->setMinMaxDistance(0.1, 0.15);
-
-      // Set hand forward direction
-      geometry_msgs::msg::Vector3Stamped vec;
-      vec.header.frame_id = hand_frame;
-      vec.vector.z = 1.0;
-      stage->setDirection(vec);
-      grasp->insert(std::move(stage));
-    }
-
-    /****************************************************
-  ---- *               Generate Grasp Pose                *
-     ***************************************************/
-    {
-      // Sample grasp pose
-      auto stage = std::make_unique<moveit::task_constructor::stages::GenerateGraspPose>("generate grasp pose");
-      stage->properties().configureInitFrom(moveit::task_constructor::Stage::PARENT);
-      stage->properties().set("marker_ns", "grasp_pose");
-      stage->setPreGraspPose("open");
-      stage->setObject("object");
-      stage->setAngleDelta(M_PI / 12);
-      stage->setMonitoredStage(current_state_ptr);  // Hook into current state
-
-//      geometry_msgs::msg::PoseStamped target_pose_msg;
-//      target_pose_msg.header.frame_id = "object";
-
-//      auto stage = std::make_unique<moveit::task_constructor::stages::GeneratePose>("generate pose");
-//      stage->properties().configureInitFrom(moveit::task_constructor::Stage::PARENT);
-//      stage->properties().set("marker_ns", "grasp_pose");
-//      stage->setPose(target_pose_msg);
-//      stage->setMonitoredStage(current_state_ptr);  // Hook into current state
-
-      Eigen::Isometry3d grasp_frame_transform;
-      Eigen::Quaterniond q = Eigen::AngleAxisd(M_PI/2, Eigen::Vector3d::UnitX())
-          * Eigen::AngleAxisd(M_PI/2, Eigen::Vector3d::UnitY())
-          * Eigen::AngleAxisd(M_PI/2, Eigen::Vector3d::UnitZ());
-      grasp_frame_transform.linear() = q.matrix();
-      grasp_frame_transform.translation().z() = 0.1;
-
-      // Compute IK
-      auto wrapper = std::make_unique<moveit::task_constructor::stages::ComputeIK>("grasp pose IK", std::move(stage));
-      wrapper->setMaxIKSolutions(8);
-      wrapper->setMinSolutionDistance(1.0);
-      wrapper->setIKFrame(grasp_frame_transform, hand_frame);
-      wrapper->properties().configureInitFrom(moveit::task_constructor::Stage::PARENT, { "eef", "group" });
-      wrapper->properties().configureInitFrom(moveit::task_constructor::Stage::INTERFACE, { "target_pose" });
-      grasp->insert(std::move(wrapper));
-
-      {
-        auto stage = std::make_unique<moveit::task_constructor::stages::ModifyPlanningScene>("allow collision (hand,object)");
-        stage->allowCollisions(
-            "object", task.getRobotModel()->getJointModelGroup(hand_group_name)->getLinkModelNamesWithCollisionGeometry(),
-            true);
-        grasp->insert(std::move(stage));
-      }
-
-      {
-        auto stage = std::make_unique<moveit::task_constructor::stages::MoveTo>("close hand", interpolation_planner);
-        stage->setGroup(hand_group_name);
-        stage->setGoal("close");
-        grasp->insert(std::move(stage));
-      }
-
-
-      {
-        auto stage = std::make_unique<moveit::task_constructor::stages::ModifyPlanningScene>("attach object");
-        stage->attachObject("object", hand_frame);
-        attach_object_stage = stage.get();
-        grasp->insert(std::move(stage));
-      }
-
-      {
-        auto stage = std::make_unique<moveit::task_constructor::stages::MoveRelative>("lift object", cartesian_planner);
-        stage->properties().configureInitFrom(moveit::task_constructor::Stage::PARENT, { "group" });
-        stage->setMinMaxDistance(0.1, 0.3);
-        stage->setIKFrame(hand_frame);
-        stage->properties().set("marker_ns", "lift_object");
-
-        // Set upward direction
-        geometry_msgs::msg::Vector3Stamped vec;
-        vec.header.frame_id = "world";
-        vec.vector.z = 1.0;
-        stage->setDirection(vec);
-        grasp->insert(std::move(stage));
-      }
-    }
-    task.add(std::move(grasp));
-  }
-  return task;
-}
-
-int main(int argc, char** argv)
-{
-//  std::this_thread::sleep_for(std::chrono::seconds(5));
-
-  rclcpp::init(argc, argv);
-
-  rclcpp::NodeOptions options;
-  options.automatically_declare_parameters_from_overrides(true);
-
-  auto mtc_task_node = std::make_shared<MTCTaskNode>(options);
-  rclcpp::executors::MultiThreadedExecutor executor;
-
-  auto spin_thread = std::make_unique<std::thread>([&executor, &mtc_task_node]() {
-    executor.add_node(mtc_task_node->getNodeBaseInterface());
-    executor.spin();
-    executor.remove_node(mtc_task_node->getNodeBaseInterface());
-  });
-
-  mtc_task_node->setupPlanningScene();
-  mtc_task_node->doTask();
-
-  executor.cancel();
-  spin_thread->join();
-
-  rclcpp::shutdown();
-  return 0;
-}
