+++ doc/tutorials/pick_and_place_with_moveit_task_constructor/package.xml
